requirements.txt
streamlit>=1.28.0
imapclient>=2.3.1
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0
